+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Prevent “warm” shutdown of Slurm workers from hanging by avoiding blocking
+  the gevent loop on worker exit.
+
 ## [1.1.2] - 2025-08-08
 
 ### Fixed